small commands refactor
