+++ bot/websocket_helper.py
@@ -1,12 +1,9 @@
 from functools import wraps
 import logging
 import random
-import re
 import time
-from typing import List
 
 from apscheduler.schedulers.background import BackgroundScheduler  # type: ignore
-from telegram import InlineKeyboardButton, InlineKeyboardMarkup
 import ujson
 import websocket  # type: ignore
 
@@ -148,40 +145,6 @@ class WebSocketHelper:
 
         self.parse_sensors(status_resp)
 
-    def _parse_custom_inline_keyboard(self, message: str):
-        def parse_button(mess: str):
-            name = re.search(r"name\s*=\s*\'(.[^\']*)\'", mess)
-            command = re.search(r"command\s*=\s*\'(.[^\']*)\'", mess)
-            if name and command:
-                gcode = "do_nothing" if command.group(1) == "delete" else f"gcode:{command.group(1)}"
-                return InlineKeyboardButton(name.group(1), callback_data=gcode)
-            else:
-                logger.warning("Bad command!")
-                return None
-
-        keyboard: List[List[InlineKeyboardButton]] = list(
-            map(
-                lambda el: list(
-                    filter(
-                        None,
-                        map(
-                            parse_button,
-                            re.findall(r"\{.[^\}]*\}", el),
-                        ),
-                    )
-                ),
-                re.findall(r"\[.[^\]]*\]", message),
-            )
-        )
-
-        title_mathc = re.search(r"message\s*=\s*\'(.[^\']*)\'", message)
-        if title_mathc:
-            title = title_mathc.group(1)
-        else:
-            title = ""
-
-        self._notifier.send_custom_inline_keyboard(title, InlineKeyboardMarkup(keyboard))
-
     def notify_gcode_reponse(self, message_params):
         if self._timelapse.manual_mode:
             if "timelapse start" in message_params:
@@ -220,7 +183,7 @@ class WebSocketHelper:
         if message_params_loc.startswith("set_notify_params "):
             self._notifier.parse_notification_params(message_params_loc)
         if message_params_loc.startswith("tgcustom_keyboard "):
-            self._parse_custom_inline_keyboard(message_params_loc)
+            self._notifier.send_custom_inline_keyboard(message_params_loc)
 
         if message_params_loc.startswith("tg_send_image"):
             self._notifier.send_image(message_params_loc)
@@ -474,7 +437,6 @@ class WebSocketHelper:
         print("lalal")
 
     def run_forever(self):
-
         # debug reasons only
         if self._log_parser:
             self.parselog()
