+++ scripts/requirements.dev.txt
@@ -1,17 +1,17 @@
 -r requirements.txt
-black==22.12.0
-isort==5.11.4
+black==23.1.0
+isort==5.12.0
 memory_profiler==0.61.0
 mypy==0.982
 numpy~=1.21.6
 opencv-python~=3.4.17.63
-pre-commit==2.21.0
-pylint==2.15.10
+pre-commit==3.0.3
+pylint==2.16.1
 pytest==7.2.1
 types-backports==0.1.3
-types-cachetools==5.2.1
+types-cachetools==5.3.0.0
 types-emoji==2.1.0.1
-types-Pillow==9.4.0.2
+types-Pillow==9.4.0.6
 types-pytz==2022.7.1.0
-types-requests==2.28.11.7
+types-requests==2.28.11.8
 types-ujson==5.7.0.0
