+++ scripts/requirements.txt
@@ -2,9 +2,9 @@ emoji==2.2.0
 Pillow==9.4.0
 PySocks==1.7.1
 python-telegram-bot==13.15
-rel==0.4.7
+rel==0.4.8
 requests==2.28.2
 tzlocal==2.1
 ujson==5.7.0
-websocket-client==1.4.2
+websocket-client==1.5.1
 wsaccel==0.6.4
